requirements.txt
Django>=3.2.0,<4.0.0
djangorestframework>=3.12.0,<4.0.0
drf-nested-routers>=0.93.0
django-cors-headers>=3.7.0
channels>=3.0.0
channels-redis>=3.3.0 